+++ appyx-components/stable/spotlight/common/src/commonMain/kotlin/com/bumble/appyx/components/spotlight/Spotlight.kt
@@ -8,9 +8,11 @@ import com.bumble.appyx.interactions.core.ui.context.TransitionBounds
 import com.bumble.appyx.interactions.core.ui.context.UiContext
 import com.bumble.appyx.interactions.core.ui.gesture.GestureFactory
 import com.bumble.appyx.interactions.core.ui.gesture.GestureSettleConfig
+import com.bumble.appyx.mapState
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.flow.StateFlow
 
 open class Spotlight<InteractionTarget : Any>(
     scope: CoroutineScope = CoroutineScope(SupervisorJob() + Dispatchers.Main),
@@ -36,4 +38,10 @@ open class Spotlight<InteractionTarget : Any>(
     gestureSettleConfig = gestureSettleConfig,
     disableAnimations = disableAnimations,
     isDebug = isDebug
-)
+) {
+    val activeIndex: StateFlow<Float> = model.output
+        .mapState(scope) { it.currentTargetState.activeIndex }
+
+    val activeElement: StateFlow<InteractionTarget?> = model.output
+        .mapState(scope) { it.currentTargetState.activeElement }
+}