+++ demos/appyx-navigation/ios/src/iosMain/kotlin/main.ios.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.window.ComposeUIViewController
 import androidx.compose.ui.zIndex
 import com.bumble.appyx.navigation.integration.IosNodeHost
 import com.bumble.appyx.navigation.integration.MainIntegrationPoint
-import com.bumble.appyx.navigation.node.container.ContainerNode
+import com.bumble.appyx.navigation.node.container.MainNavNode
 import com.bumble.appyx.navigation.ui.AppyxSampleAppTheme
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.channels.Channel
@@ -47,7 +47,7 @@ fun MainViewController() = ComposeUIViewController {
                         onBackPressedEvents = backEvents.receiveAsFlow(),
                         integrationPoint = remember { integrationPoint }
                     ) { buildContext ->
-                        ContainerNode(
+                        MainNavNode(
                             buildContext = buildContext,
                         )
                     }